requirements.txt
# Local package
-e .

# Core dependencies
numpy>=1.20
numba>=0.51
scipy>=1.7
pandas>=1.3
NEMtropy>=2.0
statsmodels>=0.12
tqdm>=4.62
matplotlib>=3.4
seaborn>=0.11
statannotations>=0.4
scikit-learn>=1.0
python-igraph>=0.9
natsort>=8.0
adjustText>=0.7
zstandard>=0.15
git+ssh://git@github.com/sztal/pathcensus.git

# Notebooks
ipython>=7.22
ipdb>=0.13
jupyterlab>=3.2
ipywidgets>=7.6
